Tehtava4/requirements.txt
openai>=1.0.0
requests>=2.31.0
beautifulsoup4>=4.12.0
python-docx>=0.8.11
PyPDF2>=3.0.0
